requirements.txt
streamlit>=1.28.0
torch>=2.0.0
transformers>=4.35.0
pillow>=9.5.0
pytesseract>=0.3.10
sentence-transformers>=2.2.0
sqlalchemy>=2.0.0
numpy>=1.24.0
psycopg2-binary>=2.9.0
scikit-learn>=1.3.0
pandas>=2.0.0
plotly>=5.15.0
python-dotenv>=1.0.0